Upload gazebo world, along with c++ plugin
feat: uploading robot model, along with building designed in the building editor took of gazebo. A gazebo world plugin c++ script is also included to welcome the user.

This commit includes the whole build my world project that has the model folder, world folder, script folder, and the CMakeLists.txt file. The model folder includes the objects and robot that go inside the gazebo world. The world file includes the building and the models. The scripts folder includes a c++ plugin script that welcomes the user to the world. The CMakeLists.txt file links the C++ code to libraries.